+++ wallet/android/app/src/main/kotlin/tech/pylons/wallet/MainActivity.kt
@@ -1,7 +1,6 @@
-package tech.pylons.flutter_wallet
+package tech.pylons.wallet
 
 import io.flutter.embedding.android.FlutterActivity
-import tech.pylons.wallet.Pigeon
 import io.flutter.embedding.engine.FlutterEngine
 
 class MainActivity: FlutterActivity() {
